DataBase complete changes

+++ src/main/java/ru/hogwarts/school/service/FacultyService.java
@@ -6,6 +6,7 @@ import ru.hogwarts.school.model.Faculty;
 import ru.hogwarts.school.repository.FacultyRepository;
 
 import java.util.Collection;
+import java.util.Optional;
 import java.util.stream.Collectors;
 
 @Service
@@ -23,26 +24,33 @@ public class FacultyService {
     }
 
     public Faculty delete(Long id) {
-        Faculty faculty = facultyRepository.getReferenceById(id);
-        facultyRepository.deleteById(id);
-        return faculty;
+        Optional<Faculty> faculty = facultyRepository.findById(id);
+        if (faculty.isPresent()) {
+            facultyRepository.deleteById(id);
+            return faculty.get();
+        }
+        return null;
     }
 
     public Faculty update(Faculty faculty) {
-        facultyRepository.save(faculty);
-        return faculty;
+        Long id = faculty.getId();
+        Optional<Faculty> faculty1 = facultyRepository.findById(id);
+        if (faculty1.isPresent()) {
+            facultyRepository.save(faculty);
+            return faculty1.get();
+        }
+        return null;
     }
 
     public Faculty read(long id) {
-            Faculty faculty = facultyRepository.getReferenceById(id);
-            return faculty;
+        Faculty faculty = facultyRepository.getReferenceById(id);
+        return faculty;
 
     }
 
-    public Collection<Faculty> getAll(){
+    public Collection<Faculty> getAll() {
         return facultyRepository.findAll();
     }
-
 
 
     public Collection<Faculty> findByColor(String color) {
